mobile/requirements.txt
toga-android>=0.5.2
python-dotenv>=1.0.0
segno>=1.5.0
aiohttp>=3.9.0
